+++ src/App.js
@@ -3,8 +3,8 @@ import React from 'react';
 
 import './App.css';
 
+
 import Form from './components/Form/Form';
-import Form2 from './components/Form2/Form2';
 
 
 
@@ -17,9 +17,9 @@ function App() {
     
     <div>
 
-      <Form />
+      
 
-      <Form2 />
+      <Form />
     </div>
