--- src/components/Form2/Form2.css
@@ -1,8 +1,0 @@
-.container{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    width: 50%;
-    margin-top: 5rem;
-}
