+++ src/components/form/Form.js
@@ -1,158 +1,42 @@
-import React, {Component} from 'react';  
-import { useSelector } from 'react-redux';
-import "./form.css"
+import React from 'react';  
+import './Form.css'
+import { useSelector,useDispatch} from 'react-redux';
+
 import Select from '../Select/Select';
-import CheckBox from '../CheckBox/CheckBox';   
-import TextArea from '../TextArea/TextArea'; 
-import Submit from '../Submit/Submit'
+import CheckBox from '../CheckBox/Checkbox';
+import { SWITCH_OPTION } from '../../redux/reducers/types';
 
+export default function Form ()  {
 
+    const conditionValue = useSelector(state => state.condition.options);
+    const operationValue = useSelector(state => state.operation);
+    const checkBoxValue = useSelector(state => state.checkbox);
 
-
-class Form extends Component {  
-  constructor(props) {
+     const dispatch = useDispatch();
     
-
-    super(props);
-
-    
-
-    this.state = {
-      newAction: {
-        condition: '',
-        operation: '',
-        check: '',
-        comment: ''
-      },
-
-      conditionOptions: ['Запущено', 'Остановлено', 'Недоступно'],
-      operationOptions: ['Старт', 'Стоп', 'Перезапуск'],
-      checkOptions: ['Снять heapdump', 'Снять threaddump']
-
-    }
-    this.handleFormSubmit = this.handleFormSubmit.bind(this);
-    this.handleInput = this.handleInput.bind(this);
-    this.handleCheckBox = this.handleCheckBox.bind(this);
-    this.handleTextArea = this.handleTextArea.bind(this);
-    this.handleFormSubmit = this.handleFormSubmit.bind(this);
-
-
-    
-  }
-
-
-
-  handleInput(e) {
-    let value = e.target.value;
-    let name = e.target.name;
-    this.setState(
-      prevState => ({
-        newAction: {
-          ...prevState.newAction,
-          [name]: value
-        }
-      }),
-      
-    );
-  }
-
-  handleTextArea(e) {
-   
-    let value = e.target.value;
-    this.setState(
-      prevState => ({
-        newAction: {
-          ...prevState.newAction,
-          comment: value
-        }
-      })
-      
-    );
-  }
-
-  handleCheckBox(e) {
-    const newSelection = e.target.value;
-    let newSelectionArray;
-
-    if (this.state.newAction.check.indexOf(newSelection) > -1) {
-      newSelectionArray = this.state.newAction.check.filter(
-        s => s !== newSelection
-      );
-    } else {
-      newSelectionArray = [...this.state.newAction.check, newSelection];
-    }
-
-    this.setState(prevState => ({
-      newAction: { ...prevState.newAction, check: newSelectionArray }
-    }));
-  }
-
-
-
-  handleFormSubmit(e) {
-    alert( this.state.newAction.operation + ' ' + this.state.newAction.comment);
-    e.preventDefault();
-  }
-
-  
-  
-  render() {
-   
-
-    return (
-      <form className="container" onSubmit={this.handleFormSubmit}>
-
-          
-
-       <Select 
-       title={'Состояние'}
-       name={'condition'}
-       options = {this.state.conditionOptions} 
-       value = {this.state.newAction.condition}
-       placeholder = {'Выберите состояние'}
-       handleChange = {this.handleInput}
-       />
-
-       <Select 
-       title={'Операция'}
-       name={'operation'}
-       options = {this.state.operationOptions} 
-       value = {this.state.newAction.operation}
-       placeholder = {'Выберите операцию'}
-       handleChange = {this.handleInput}
-       />
-       
-       <TextArea 
-        title={"Комментарий"}
-        value={this.state.newAction.comment}
-        name={"comment"}
-        rows={5}
-        handleChange={this.handleTextArea}
-        placeholder={"Введите комментарий"}
-        /> 
-
-
+    return(
+        <div className='container'>
+        <Select
+            title={'Состояние'}
+            name={'condition'}
+            options = {conditionValue} 
+            value = {conditionValue}
+            onChange = {() => dispatch(SWITCH_OPTION())}
+        />
+        <Select 
+            title={'Операция'}
+            name={'operation'}
+            options = {operationValue} 
+            value = {operationValue}
+            onChange = {() => dispatch(SWITCH_OPTION())}
+        />
         <CheckBox
-        title={""}
-        name={"check"}
-        options={this.state.checkOptions}
-        selectedOptions={this.state.newAction.check}
-        handleChange={this.handleCheckBox} /> 
-        
-        <Submit 
-        action={this.handleFormSubmit}
-        
-        title={"Выполнить"}
-        
-        
-        /> 
-        
-      </form>
-      
-    );
-    
-  }
-  
-}
+            title={""}
+            name={"check"}
+            options={checkBoxValue}
+            selectedOptions={checkBoxValue}
+        />
 
-export default Form;
+        </div>
+    )
+}
